refactor(charts): share PolishedChartTooltip across cartesian and pie (ENG-1140)
Extract the polished hover tooltip from cartesian-chart.tsx into its own
polished-tooltip.tsx so the pie chart can use it too. The pie was still
falling back to the shadcn ChartTooltipContent, which gave a different
visual treatment on hover than the surrounding bar / line / area charts.

PolishedChartTooltip now handles both call patterns:
- Cartesian: Recharts passes the x-axis value as `label`, used as the
  card header (date / category).
- Pie: Recharts leaves `label` empty and exposes the slice name on
  payload[0].name; fall back to that so the donut tooltip header reads
  e.g. "Founder" or "Jun 2, 2026" cleanly.

Drops the now-dead PieTooltipRow / pieTooltipFormatter / ChartTooltipContent
import — one tooltip implementation for every chart type.
