+++ apps/web/modules/ee/analysis/charts/components/cartesian-chart.tsx
@@ -1,64 +1,12 @@
 "use client";
 
 import { type ElementType, type ReactNode } from "react";
-import { useTranslation } from "react-i18next";
 import { CartesianGrid, XAxis, YAxis } from "recharts";
-import {
-  CHART_BRAND_DARK,
-  formatCellValue,
-  formatXAxisTick,
-} from "@/modules/ee/analysis/charts/lib/chart-utils";
-import { formatCubeColumnHeader } from "@/modules/ee/analysis/lib/schema-definition";
+import { formatXAxisTick } from "@/modules/ee/analysis/charts/lib/chart-utils";
 import type { TChartDataRow } from "@/modules/ee/analysis/types/analysis";
 import type { ChartConfig } from "@/modules/ui/components/chart";
 import { ChartContainer, ChartLegend, ChartLegendContent, ChartTooltip } from "@/modules/ui/components/chart";
-
-type TooltipPayloadItem = {
-  dataKey?: string;
-  name?: string | number;
-  value?: unknown;
-  color?: string;
-  payload?: { fill?: string };
-};
-
-type RechartsTooltipProps = {
-  active?: boolean;
-  payload?: TooltipPayloadItem[];
-  label?: string | number;
-};
-
-/**
- * Polished hover tooltip card matching the Twenty-style target: prominent
- * date header, indicator + measure + value per series, comfortable padding
- * so values are easy to read without scanning a tiny black box.
- */
-const PolishedChartTooltip = ({ active, payload, label }: Readonly<RechartsTooltipProps>) => {
-  const { t } = useTranslation();
-  if (!active || !payload?.length) return null;
-
-  return (
-    <div className="min-w-[180px] rounded-lg border border-slate-200 bg-white px-3 py-2.5 shadow-lg">
-      <div className="mb-2 text-sm font-medium text-slate-800">{formatXAxisTick(label)}</div>
-      <div className="flex flex-col gap-1.5">
-        {payload.map((item) => {
-          const key = item.dataKey ?? String(item.name ?? "");
-          const indicatorColor = item.color ?? item.payload?.fill ?? CHART_BRAND_DARK;
-          return (
-            <div key={key} className="flex items-center justify-between gap-3">
-              <div className="flex items-center gap-2">
-                <div className="size-2.5 shrink-0 rounded-full" style={{ backgroundColor: indicatorColor }} />
-                <span className="text-sm text-slate-500">{formatCubeColumnHeader(key, t)}</span>
-              </div>
-              <span className="text-foreground text-sm font-medium tabular-nums">
-                {formatCellValue(item.value)}
-              </span>
-            </div>
-          );
-        })}
-      </div>
-    </div>
-  );
-};
+import { PolishedChartTooltip } from "./polished-tooltip";
 
 export interface CartesianChartProps {
   data: TChartDataRow[];
